add enum parsing with libclang

+++ kclang/src/main/kotlin/klang/domain/NativeEnumeration.kt
@@ -2,5 +2,5 @@ package klang.domain
 
 data class NativeEnumeration(
 	val name: String,
-	val values: List<Pair<String, Int>> = emptyList()
+	val values: List<Pair<String, Long>> = emptyList()
 ) : NativeDeclaration
